backend/requirements.txt
# =============================================================================
# MIYA AI ASSISTANT - Python Dependencies
# =============================================================================

# ── Core Framework ───────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── LLM Engine ───────────────────────────────────────────────────────────
llama-cpp-python>=0.2.50

# ── Async Utilities ──────────────────────────────────────────────────────
aiohttp>=3.9.0
aiofiles>=23.2.0
httpx>=0.25.0

# ── Vector Database ──────────────────────────────────────────────────────
chromadb>=0.4.15
sentence-transformers>=2.2.0

# ── Structured Database ──────────────────────────────────────────────────
aiosqlite>=0.19.0

# ── Cache ────────────────────────────────────────────────────────────────
redis[hiredis]>=5.0.0

# ── Object Storage ───────────────────────────────────────────────────────
minio>=7.2.0

# ── Search ───────────────────────────────────────────────────────────────
duckduckgo-search>=4.0

# ── Browser Automation ───────────────────────────────────────────────────
playwright>=1.40.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# ── Media Processing ────────────────────────────────────────────────────
faster-whisper>=0.10.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── Code Quality ─────────────────────────────────────────────────────────
ruff>=0.1.0

# ── Data Processing ──────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# ── Container Management ────────────────────────────────────────────────
docker>=6.1.0

# ── SSH ──────────────────────────────────────────────────────────────────
asyncssh>=2.14.0

# ── Authentication ───────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ── Logging ──────────────────────────────────────────────────────────────
structlog>=23.2.0
loguru>=0.7.0

# ── Monitoring ───────────────────────────────────────────────────────────
prometheus-client>=0.19.0

# ── Desktop UI ───────────────────────────────────────────────────────────
gradio>=4.8.0

# ── Fine-Tuning (Unsloth + QLoRA) ────────────────────────────────────────
unsloth
peft>=0.7.0
trl>=0.7.0
datasets>=2.16.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# ── Testing ──────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ── Typing / Linting ────────────────────────────────────────────────────
mypy>=1.7.0
